[skip ci] Refactor
