+++ packages/math/line.test.ts
@@ -1,6 +1,5 @@
-import { line, lineSegmentIntersectionPoints, linesIntersectAt } from "./line";
+import { line, linesIntersectAt } from "./line";
 import { pointFrom } from "./point";
-import { lineSegment } from "./segment";
 
 describe("line-line intersections", () => {
   it("should correctly detect intersection at origin", () => {
@@ -28,24 +27,5 @@ describe("line-line intersections", () => {
         line(pointFrom(10, 0), pointFrom(10, 10)),
       ),
     ).toBe(null);
-  });
-});
-
-describe("line-segment intersections", () => {
-  it("should correctly detect intersection", () => {
-    expect(
-      lineSegmentIntersectionPoints(
-        lineSegment(pointFrom(0, 0), pointFrom(5, 0)),
-        lineSegment(pointFrom(2, -2), pointFrom(3, 2)),
-      ),
-    ).toEqual(pointFrom(2.5, 0));
-  });
-  it("should correctly detect non-intersection", () => {
-    expect(
-      lineSegmentIntersectionPoints(
-        lineSegment(pointFrom(0, 0), pointFrom(5, 0)),
-        lineSegment(pointFrom(3, 1), pointFrom(4, 4)),
-      ),
-    ).toEqual(null);
   });
 });
