+++ packages/math/segment.test.ts
@@ -1,0 +1,21 @@
+import { pointFrom } from "./point";
+import { lineSegment, lineSegmentIntersectionPoints } from "./segment";
+
+describe("line-segment intersections", () => {
+  it("should correctly detect intersection", () => {
+    expect(
+      lineSegmentIntersectionPoints(
+        lineSegment(pointFrom(0, 0), pointFrom(5, 0)),
+        lineSegment(pointFrom(2, -2), pointFrom(3, 2)),
+      ),
+    ).toEqual(pointFrom(2.5, 0));
+  });
+  it("should correctly detect non-intersection", () => {
+    expect(
+      lineSegmentIntersectionPoints(
+        lineSegment(pointFrom(0, 0), pointFrom(5, 0)),
+        lineSegment(pointFrom(3, 1), pointFrom(4, 4)),
+      ),
+    ).toEqual(null);
+  });
+});
